separate the index route file and make new one named product-admin
