+++ routes/index.js
@@ -46,21 +46,4 @@ router.get('/products/:urlPath', function(req, res, next) {
     dbClient.getProducts({ urlPath }, callBack);
 });
 
-router.get('/products-admin/add', (req, res) => {
-    res.render("add-product");
-})
-
-
-router.post('/products-admin', (req, res) => {
-    const callBack = (err) => {
-        if (err) {
-            res.sendStatus(500)
-        } else {
-            res.redirect('/');
-        }
-    }
-    const query = req.body
-    dbClient.addProducts(query, callBack);
-})
-
 module.exports = router;
